backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Authentication
PyJWT==2.8.0

# Async support
httpx==0.26.0
aiofiles==23.2.1

# Redis queue
redis==5.0.1
rq==1.16.0

# Video processing
scenedetect[opencv]==0.6.2
opencv-python-headless==4.9.0.80
moviepy==1.0.3
ffmpeg-python==0.2.0

# MinIO storage
minio==7.2.3

# ElevenLabs TTS (using direct HTTP requests, no SDK needed)
# elevenlabs==1.1.2  # Removed - using requests instead for compatibility

# Configuration
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Utilities
Pillow==10.2.0
numpy==1.26.3
thefuzz==0.22.1
python-levenshtein==0.25.1

# Google Gemini API
google-generativeai>=0.8.0

# Vector embeddings and semantic matching
sentence-transformers>=2.2.0
spacy>=3.7.0
scikit-learn>=1.3.0